skin_lesion_classification/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
albumentations>=1.4.0
numpy>=1.24.0
Pillow>=10.0.0
pydantic>=2.0.0
